chore: split price check and cf check
