+++ .github/workflows/hourly.yml
@@ -81,6 +81,8 @@ jobs:
         run: python -m pendle.main
       - name: Run LRT Peg Checker Balancer
         run: python -m lrt-pegs.balancer.main
+      - name: Run USD0 Peg Price Checker
+        run: python -m usd0.price
 
       # save file to cache
       - name: Cache cache-id.txt
